refactor(hub): extraer HubView del overlay para reusar como workspace
